backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
tortoise-orm==0.21.0
aerich==0.7.2
aiomysql==0.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.8.0
pydantic-settings==2.4.0
python-dotenv==1.0.1
